Get a set of feature data for an area or planning grid

This is a wrapper of `get_bathymetry()`, `get_seamounts_buffered()`, `get_knolls()`, `get_geomorphology()`, `get_coral_habitat()`, and `get_enviro_regions()`. See the indvidual functions for details.

Usage

get_features(
  area_polygon = NULL,
  spatial_grid = NULL,
  features = c("bathymetry", "seamounts", "knolls", "geomorphology", "corals",
    "enviro_regions"),
  seamount_buffer = 30000,
  bathy_resolution = 1,
  antipatharia_threshold = 22,
  octocoral_threshold = 2,
  enviro_clusters = NULL,
  max_enviro_clusters = 6,
  antimeridian = NULL
)

Arguments

area_polygon

an `sf` polygon or multipolygon object of the area of interest (e.g., a country's EEZ)

spatial_grid

a `terra::rast()` or `sf` planning grid with the desired resolution and coordinate reference system generated by `get_grid()`. For a raster, values in the planning grid cells should be 1, while all other values are NA

features

a vector of feature names, can include: "bathymetry", "seamounts", "knolls", "geomorphology", "corals", "enviro_regions"

seamount_buffer

`numeric`; the distance from the seamount peak to include in the output. Distance should be in the same units as the area_polygon or spatial_grid provided, use e.g. `sf::st_crs(spatial_grid, parameters = TRUE)$units_gdal` to check what units your planning grid or area polygon is in (works for raster as well as sf objects)

bathy_resolution

`numeric`; the resolution (in minutes) of data to pull from the ETOPO 2022 Global Relief model. Values less than 1 can only be 0.5 (30 arc seconds) and 0.25 (15 arc seconds)

antipatharia_threshold

`numeric` between 0 and 100; the threshold value for habitat suitability for antipatharia corals to be considered present (default is 22, as defined in Yesson et al., 2017)

octocoral_threshold

`numeric` between 0 and 7; the threshold value for how many species (of 7) should be predicted present in an area for octocorals to be considered present (default is 2)

enviro_clusters

`numeric`; the number of environmental regions to cluster the data into - to be used when a clustering algorithm is not necessary (default is NULL)

max_enviro_clusters

`numeric`; the maximum number of environmental regions to try when using the clustering algorithm (default is 8)

antimeridian

Does `area_polygon` or `spatial_grid` span the antimeridian? If so, this should be set to `TRUE`, otherwise set to `FALSE`. If set to `NULL` (default) the function will try to check if data spans the antimeridian and set this appropriately.

Value

If an `area_polygon` is supplied, a list of feature data is returned (will be mixed raster and sf objects). If a `spatial_grid` is supplied, a raster stack or `sf` of gridded data is returned, depending on `spatial_grid` format.

Examples

# Grab EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
#> Cache is fresh. Reading: /tmp/RtmpwMyPbE/eez-2205f12f/eez.shp
#> (Last Modified: 2024-02-09 06:43:34.396345)
# Get a feature data for Bermuda
feature_set <- get_features(area_polygon = bermuda_eez)
#> Getting depth zones...
#> This may take seconds to minutes, depending on grid size
#> Getting seamount data...
#> Spherical geometry (s2) switched off
#> although coordinates are longitude/latitude, st_intersection assumes that they
#> are planar
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Warning: st_buffer does not correctly buffer longitude/latitude data
#> dist is assumed to be in decimal degrees (arc_degrees).
#> Warning: bounding box has potentially an invalid value range for longlat data
#> although coordinates are longitude/latitude, st_union assumes that they are
#> planar
#> Spherical geometry (s2) switched on
#> Getting knoll data...
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Getting geomorphology data...
#> Getting coral data...
#> |-- Coral data found for antipatharia, cold corals, octocorals
#> Getting environmental regions data... This could take several minutes
# Get feature data in a planning grid
planning_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
# Grab all the data
features_gridded <- get_features(spatial_grid = planning_grid)
#> Getting depth zones...
#> This may take seconds to minutes, depending on grid size
#> Getting seamount data...
#> Spherical geometry (s2) switched off
#> although coordinates are longitude/latitude, st_intersection assumes that they
#> are planar
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Spherical geometry (s2) switched on
#> Getting knoll data...
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Getting geomorphology data...
#> Getting coral data...
#> |-- Coral data found for antipatharia coral, cold coral, octocoral
#> Getting environmental regions data... This could take several minutes