Get geomorphology for an area of interest or planning grid

This function gets geomorpholgical data within an area polygon or planning grid depending on which is supplied

Usage

get_geomorphology(
  area_polygon = NULL,
  spatial_grid = NULL,
  antimeridian = NULL
)

Arguments

area_polygon

an `sf` polygon or multipolygon object of the area of interest (e.g., a country's EEZ)

spatial_grid

a `terra::rast()` or `sf` planning grid with the desired resolution and coordinate reference system generated by `get_grid()`. For a raster, values in the planning grid cells should be 1, while all other values are NA

antimeridian

Does `area_polygon` or `spatial_grid` span the antimeridian? If so, this should be set to `TRUE`, otherwise set to `FALSE`. If set to `NULL` (default) the function will try to check if data spans the antimeridian and set this appropriately.

Value

If an `area_polygon` is supplied, an `sf` object with each row a different geomorphological feature. If a `spatial_grid` is supplied, a raster stack of geomorphological features or `sf` object with a geomorphological feature in each column is returned, depending on `spatial_grid` format.

Details

Geomorphological features are from the [Harris et al. 2014](https://doi.org/10.1016/j.margeo.2014.01.011) dataset, available at: https://www.bluehabitats.org. The features included are limited to those listed in [Cecarelli et al. 2021](https://doi.org/10.3389/fmars.2021.634574) Table 5.

Examples

# Grab EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
#> Cache is fresh. Reading: /tmp/RtmpxLQp9P/eez-2205f12f/eez.shp
#> (Last Modified: 2024-02-08 05:00:36.60504)
# Get geomorphology for the EEZ
geomorphology <- get_geomorphology(area_polygon = bermuda_eez)
# Get geomorphological features in spatial_grid
planning_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
geomorph_gridded <- get_geomorphology(spatial_grid = planning_grid)