Get seamount peaks for area of interest or planning grid

This function gets seamount peaks within an area polygon or planning grid depending on which is supplied

Usage

get_seamount_peaks(
  area_polygon = NULL,
  spatial_grid = NULL,
  name = "seamounts",
  antimeridian = NULL
)

Arguments

area_polygon

an `sf` polygon or multipolygon object of the area of interest (e.g., a country's EEZ)

spatial_grid

a `terra::rast()` or `sf` planning grid with the desired resolution and coordinate reference system generated by `get_grid()`. For a raster, values in the planning grid cells should be 1, while all other values are NA

name

name to give to the raster or sf column

antimeridian

Does `area_polygon` or `spatial_grid` span the antimeridian? If so, this should be set to `TRUE`, otherwise set to `FALSE`. If set to `NULL` (default) the function will try to check if data spans the antimeridian and set this appropriately.

Value

If an `area_polygon` is supplied, an `sf` seamounts object. If a `spatial_grid` is supplied, a raster or `sf` of gridded seamount data, depending on `spatial_grid` format.

Details

Seamounts are classified as peaks at least 1000m higher than the surrounding seafloor [Morato et al. 2008](https://doi.org/10.3354/meps07268). The seamounts peak dataset is from [Yeson et al. 2021](https://doi.org/10.14324/111.444/ucloe.000030).

Examples

# Get EEZ data first 
bermuda_eez <- get_area(area_name = "Bermuda", mregions_column = "territory1")
#> Cache is fresh. Reading: /tmp/RtmpLoPEmn/eez-2205f12f/eez.shp
#> (Last Modified: 2024-02-09 00:40:19.24325)
# Get seamount peaks
seamount_peaks <- get_seamount_peaks(area_polygon = bermuda_eez)
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
# Get seamounts in a planning grid
planning_grid <- get_grid(area_polygon = bermuda_eez, projection_crs = '+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs', resolution = 5000)
seamounts_gridded <- get_seamount_peaks(spatial_grid = planning_grid)
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries